fixing babel parsing
